Small refactoring and Group entity
